moved campaigns data to Campaigns Page and started working on filter function
