--- src/App.js
@@ -1,11 +1,0 @@
-import React from "react";
-import Table from "./Components/Campaigns/Table";
-
-
-function App() {
-      return (
-        <Table />
-      );
-}
-
-export default App;
